+++ hydromt/models/__init__.py
@@ -2,6 +2,7 @@
 """HydroMT models API."""
 from hydromt._compat import HAS_XUGRID
 from hydromt.models.api import Model
+from hydromt.models.components import ModelComponent, ModelRegionComponent
 from hydromt.models.components.grid import GridModel
 from hydromt.models.components.network import NetworkModel
 from hydromt.models.components.vector import VectorModel
@@ -9,6 +10,8 @@ from hydromt.models.plugins import ModelCatalog
 
 if HAS_XUGRID:
     from hydromt.models.components.mesh import MeshModel
+
+from hydromt.models.root import ModeLike, ModelMode, ModelRoot
 
 # expose global MODELS object which discovers and loads
 # any local generalized or plugin model class on-the-fly