+++ tests/metadata_resolvers/test_convention_resolver.py
@@ -1,3 +1,4 @@
+import pytest
 from polyfactory.factories.pydantic_factory import ModelFactory
 from polyfactory.pytest_plugin import register_fixture
 
@@ -13,6 +14,7 @@ class DataSourceFactory(ModelFactory[DataSource]):
     metadata_resolver = ConventionResolver()
 
 
+@pytest.mark.skip("see ticket #805 ")
 class TestConventionResolver:
     def test_resolve(self, data_source_factory: DataSourceFactory):
         mock_source = data_source_factory.build()