+++ tests/workflows/test_basin_mask.py
@@ -1,139 +1,16 @@
 # -*- coding: utf-8 -*-
 """Tests for the hydromt.workflows.basin_mask."""
 import logging
-import os
 
-import geopandas as gpd
 import numpy as np
 import pytest
-import xarray as xr
 
 import hydromt
-from hydromt import DataCatalog, raster
-from hydromt.models import MODELS
 from hydromt.workflows.basin_mask import (
-    _check_size,
-    _parse_region_value,
     get_basin_geometry,
-    parse_region,
 )
 
-logger = logging.getLogger("tets_basin")
-
-
-@pytest.mark.skip(reason="Needs refactor to Pydantic BaseModel.")
-def test_region(tmpdir, world, geodf, rioda):
-    # prepare test data
-    fn_gdf = str(tmpdir.join("world.geojson"))
-    world.to_file(fn_gdf, driver="GeoJSON")
-    fn_grid = str(tmpdir.join("grid.tif"))
-    rioda.raster.to_raster(fn_grid)
-    cat = DataCatalog()
-    cat.from_dict(
-        {
-            "world": {
-                "path": fn_gdf,
-                "data_type": "GeoDataFrame",
-                "driver": "vector",
-            },
-            "grid": {
-                "path": fn_grid,
-                "data_type": "RasterDataset",
-                "driver": "raster",
-            },
-        }
-    )
-
-    # model
-    region = {"region": [0.0, -1.0]}
-    with pytest.raises(ValueError, match=r"Region key .* not understood.*"):
-        parse_region(region)
-
-    model = MODELS.generic[0]
-    root = str(tmpdir.join(model)) + "_test_region"
-    if not os.path.isdir(root):
-        os.mkdir(root)
-    region = {model: root}
-    kind, region = parse_region(region)
-    assert kind == "model"
-
-    # geom
-    kind, region = parse_region({"geom": world})
-    assert isinstance(region["geom"], gpd.GeoDataFrame)
-    kind, region = parse_region({"geom": fn_gdf})
-    assert isinstance(region["geom"], gpd.GeoDataFrame)
-    kind, region = parse_region({"geom": "world"}, data_catalog=cat)
-    assert isinstance(region["geom"], gpd.GeoDataFrame)
-    # geom:  points should fail
-    region = {"geom": geodf}
-    with pytest.raises(ValueError, match=r"Region value.*"):
-        kind, region = parse_region(region)
-
-    # grid
-    kind, region = parse_region({"grid": rioda})
-    assert isinstance(region["grid"], xr.DataArray)
-    kind, region = parse_region({"grid": fn_grid})
-    assert isinstance(region["grid"], xr.DataArray)
-    kind, region = parse_region({"grid": "grid"}, data_catalog=cat)
-    assert isinstance(region["grid"], xr.DataArray)
-
-    # basid
-    region = {"basin": [1001, 1002, 1003, 1004, 1005]}
-    kind, region = parse_region(region)
-    assert kind == "basin"
-    assert region.get("basid") == [1001, 1002, 1003, 1004, 1005]
-    region = {"basin": 101}
-    kind, region = parse_region(region)
-    assert kind == "basin"
-    assert region.get("basid") == 101
-
-    # bbox
-    region = {"outlet": [0.0, -5.0, 3.0, 0.0]}
-    kind, region = parse_region(region)
-    assert kind == "outlet"
-    assert "bbox" in region
-
-    # xy
-    region = {"subbasin": [1.0, -1.0], "uparea": 5.0, "bounds": [0.0, -5.0, 3.0, 0.0]}
-    kind, region = parse_region(region)
-    assert kind == "subbasin"
-    assert "xy" in region
-    assert "bounds" in region
-    region = {"basin": [[1.0, 1.5], [0.0, -1.0]]}
-    kind, region = parse_region(region)
-    assert "xy" in region
-    region = {"interbasin": geodf}
-    kind, region = parse_region(region)
-    assert "xy" in region
-
-
-def test_region_value():
-    data_catalog = DataCatalog()
-    array = np.array([1001, 1002, 1003, 1004, 1005])
-    kwarg = _parse_region_value(array, data_catalog=data_catalog)
-    assert kwarg.get("basid") == array.tolist()
-    xy = (1.0, -1.0)
-    kwarg = _parse_region_value(xy, data_catalog=data_catalog)
-    assert kwarg.get("xy") == xy
-    root = "./"
-    kwarg = _parse_region_value(root, data_catalog=data_catalog)
-    assert kwarg.get("root") == root
-
-
-def test_check_size(caplog):
-    test_raster = raster.full_from_transform(
-        transform=[0.5, 0.0, 3.0, 0.0, -0.5, -9.0],
-        shape=(13000, 13000),
-        nodata=-1,
-        name="test",
-        crs=4326,
-        lazy=True,  # create lazy dask array instead of numpy array
-    )
-    _check_size(test_raster)
-    assert (
-        "Loading very large spatial domain to derive a subbasin. "
-        "Provide initial 'bounds' if this takes too long." in caplog.text
-    )
+logger = logging.getLogger(__name__)
 
 
 @pytest.mark.skip(reason="Needs implementation of RasterDataSet.")